Removed comment about what day the week starts on
I asked about the option in the mobile app for what day the week starts on:
"1. I don't see this option anywhere in the API, anyone know if it's there?
2. Anyone know what this is used for? I thought maybe for determining future due dates for dailies that recur weekly."

@alys replied:
"It's purely for displaying date pickers in calendar format. It's only on the mobile apps because the date picker currently used in the website doesn't have an option to change which day is shown first.

I believe the setting is stored within the app data on your mobile device (not within your Habitica user account's data which is stored on Habitica's database server). This means it wouldn't be mentioned anywhere in the API and wouldn't be readable or writable using the API."

So it has nothing to do with when dailies that recur weekly are due. I think I can safely tell rrule that the week starts on Sunday.